Changed defer checks to use explicit `is True`.
Using `if defer is True:` for checks instead of `if defer:` catches a
lot of weird errors that can occur if the positional args to any of the
reflected planning-method functions are shifted by one.

The previous check would return a Future if an extra argument got
passed, which concealed exceptions indicating that the arguments made
no sense, and would be passed to subsequent code until something
actually tried to query a Trajectory method on the Future.