Changed default 'execute' behavior to False.
Currently, it is very scary that calling a Plan...() function
anywhere in the code without 'execute=False' can cause a robot
to start moving and doing things.

It does not seem at all reasonable that this is the default behavior,
especially now that we are increasingly using pipelined planning calls
and post processing trajectories heavily before execution in demos
--where execute=True is exactly the wrong thing to do.

We should not have to pass a non-default argument to every single
planning call to keep the robot from moving.