Added a check within snap planner for one-waypoint trajectories.
If a plan is requested that has the same start and goal configuration,
SnapPlanner previously returned a two-waypoint trajectory with
the same point as the start and goal.

This leads to numerical instability and uniqueness problems, because
there is no reasonable way to interpolate between the two waypoints,
since they occur at the same time and have a delta-time of 0.0.

This fix explicitly checks this condition, and only returns a
two-waypoint trajectory if start and goal are different.