Make `Cloned()` handle `None`.
Currently, `Cloned(None)` will throw the exception that "`None` is not in the cloned environment".  This is a bit annoying in cases where you to clone an optional object reference, such as a variable that is either a KinBody or `None`.

This change adds logic to make `Cloned(None) = None`.