Added a helper function to extract joint groups.
This commit creates a helper function that copies OpenRAVE's
hardcoded table of joint derivative names and adds an accessor
function for retrieving them from a configuration specification.

This allows users to directly query the n-th derivative group of
a trajectory without having to iterate through other derivatives.

I could not think of a cleaner way to do this, but at least this
avoid having the values hard-coded in multiple places.

Note that this function only handles joint value derivatives.  A
corresponding function for affine derivatives could be added in the
future.