Added tolerances parameters to SimplifyTrajectory.
Instead of simply using DOF resolution, this allows SimplifyTrajectory
to take a list of DOF tolerances.  This is useful in situations where
it is desirable to force a more accurate simplification.

This also forwards the tolerance argument through the default OpenRAVE
retimer in `prpy.planning.retimer`, since it includes a SimplifyTrajectory
call by default.