Removes the ill-fated Trollius support from PlanningMethod.
Trollius support never actually worked very well (or at all), and
with the addition of `prpy.futures` we have a much better interface
for doing parallel operations.

This removes the `defer` options on all prpy planning methods, and
replaced the `Ranked` implementation with one that uses `prpy.futures`
and plain threading instead.