Fixed spurious CloneException errors in planning stack.
Due to OpenRAVE weirdness, we are forced to throw a CloneException
if a robot is cloned from a self-colliding configuration.  We also
clone at the beginning of our planning operations.

This means that calling a planning method while a robot is in self
collision would create a CloneException, which is generally not
expected or handled in code that is calling a planning method.  It
is also not clear to calling code that this is actually a failure
within a planner, and it breaks MetaPlanner logic.

In this PR, ClonedPlanningMethods will catch CloneExceptions and
wrap them as ClonedPlanningErrors instead, which inherit from
PlanningError and are thus handled correctly throughout the
planning stack as a form of planning failure.