Adds goal endpoint checking to snap planner.
As per discussion in #297, it is valuable to explicitly check the
goal state of a planning operation.

This PR addresses the change in SnapPlanner by modifying the
underlying VanDerCorputSampleGenerator to always return the start
and end points. There doesn't seem to be a compelling reason for
this not to be the case, as the generated samples are aliased over
the collision checking interval in an arbitrary fashion anyway.