Changes VanDerCorputSampleGenerator to include goal.
This refactors VanDerCorputSampleGenerator to include the start
and end (goal) values as the first two samples to be checked,
followed by the discretized intermediate checks within the
interval range.

This is an alternate way to address #297 and #303, but avoids
the non-termination problem introduced in #303 by simplifying
the generator loop.  The resulting output is identical to the
expected output of #303.